// src/api/model/invoiceVatClaimDtoClaimType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

/**
 * FULL = 100% deduction, PARTIAL = ratio in (0,1). Required when shouldClaimVat=true.
 */
export type InvoiceVatClaimDtoClaimType = typeof InvoiceVatClaimDtoClaimType[keyof typeof InvoiceVatClaimDtoClaimType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceVatClaimDtoClaimType = {
  FULL: 'FULL',
  PARTIAL: 'PARTIAL',
} as const;